pyproject.toml
[project]
name = "coreascher"
version = "0.1.0"
description = "基于 CrewAI 的文献综述助手"
authors = [
    { name = "YiCheng", email = "yicheng@example.com" }
]
dependencies = [
    "crewai>=0.95.0",
    "langchain>=0.1.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "loguru>=0.7.2",
    "PyYAML>=6.0.1"
]
requires-python = ">=3.10"
readme = "README.md"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
include = ["coreascher*"]
namespaces = false
